changed patron blocks
